Moved NumberExtensionsTest from itest to pure unit test and merged redundant calsses
